Clean up entry point
