+++ public/js/UnleashApp.jsx
@@ -1,19 +1,10 @@
 var React                   = require('react');
-var UserStore               = require('./stores/UserStore');
+var Router                  = require('react-router');
 var Menu                    = require('./components/Menu');
 var ErrorMessages           = require('./components/ErrorMessages');
 var initalizer              = require('./stores/initalizer');
-var LogEntriesComponent     = require('./components/log/LogEntriesComponent');
-var FeatureTogglesComponent = require('./components/feature/FeatureTogglesComponent');
-var StrategiesComponent     = require('./components/strategy/StrategiesComponent');
-var ArchiveFeatureComponent = require('./components/feature/ArchiveFeatureComponent');
-var Router                  = require('react-router');
-var DefaultRoute = Router.DefaultRoute;
 var Link = Router.Link;
-var Route = Router.Route;
 var RouteHandler = Router.RouteHandler;
-
-UserStore.init();
 
 var UnleashApp = React.createClass({
     contextTypes: {
@@ -24,22 +15,22 @@ var UnleashApp = React.createClass({
         initalizer(30);
     },
 
+    renderLink: function(id, label) {
+        return    (
+            <Link to={id} className="nav-element centerify" activeClassName="nav-active">
+                <span className="topbar-nav-svg-caption caption showbydefault no-break">{label}</span>
+            </Link>
+        );
+    },
+
     render: function () {
         return (
             <div>
                 <Menu>
-                    <Link to="features" className="nav-element centerify" activeClassName="nav-active">
-                        <span className="topbar-nav-svg-caption caption showbydefault no-break">Features</span>
-                    </Link>
-                    <Link to="strategies" className="nav-element centerify" activeClassName="nav-active">
-                        <span className="topbar-nav-svg-caption caption showbydefault no-break">Strategies</span>
-                    </Link>
-                    <Link to="log" className="nav-element centerify" activeClassName="nav-active">
-                        <span className="topbar-nav-svg-caption caption showbydefault no-break">Log</span>
-                    </Link>
-                    <Link to="archive" className="nav-element centerify" activeClassName="nav-active">
-                        <span className="topbar-nav-svg-caption caption showbydefault no-break">Archive</span>
-                    </Link>
+                    {this.renderLink("features",    "Toggles")}
+                    {this.renderLink("strategies",  "Strategies")}
+                    {this.renderLink("log",         "Log")}
+                    {this.renderLink("archive",     "Archive")}
                 </Menu>
                 <div className="container">
                     <div className="page">
@@ -56,19 +47,6 @@ var UnleashApp = React.createClass({
             </div>
         );
     }
-});
-
-var routes = (
-    <Route name="app" path="/" handler={UnleashApp}>
-        <Route name="strategies" handler={StrategiesComponent}/>
-        <Route name="log" handler={LogEntriesComponent}/>
-        <Route name="archive" handler={ArchiveFeatureComponent}/>
-        <DefaultRoute name="features" handler={FeatureTogglesComponent}/>
-    </Route>
-);
-
-Router.run(routes, function (Handler) {
-    React.render(<Handler/>, document.getElementById('content'));
 });
